refactor: 파일 분리 및 ui 모델 프로퍼티 수정
