+++ android/app/src/main/java/poke/rogue/helper/presentation/type/model/MatchedTypesUiModel.kt
@@ -1,12 +1,24 @@
 package poke.rogue.helper.presentation.type.model
 
-import androidx.annotation.ColorRes
-import androidx.annotation.StringRes
+import poke.rogue.helper.data.model.MatchedTypes
 
 data class MatchedTypesUiModel(
     val selectedType: TypeUiModel,
     val isMyType: Boolean,
-    @StringRes val matchedResultDescriptionResId: Int,
-    @ColorRes val matchedResultColorResId: Int,
+    val matchedResultUi: MatchedResultUiModel,
     val matchedItem: List<TypeUiModel>,
 )
+
+fun MatchedTypes.toUi(
+    typeId: Int,
+    isMyType: Boolean,
+): MatchedTypesUiModel {
+    val inputTypeUi = TypeUiModel.fromId(typeId) ?: throw IllegalArgumentException("Unknown type ID: $typeId")
+    val matchedResultUi = MatchedResultUiModel.fromMatchedResult(this.matchedResult)
+    return MatchedTypesUiModel(
+        selectedType = inputTypeUi,
+        isMyType = isMyType,
+        matchedResultUi = matchedResultUi,
+        matchedItem = this.types.map { it.toUi() },
+    )
+}
